+++ docs/make.jl
@@ -30,6 +30,7 @@ makedocs(;
             "manual/queries.md",
             "manual/systems.md",
             "manual/resources.md",
+            "manual/relations.md",
             "manual/events.md",
             "manual/architecture.md"
         ],